{"schema":"decodex/commit/1","summary":"Split structured field validation modules","authority":"manual"}
